release 1.4.0: A 端支持四种架构（arm64-v8a / armeabi-v7a / x86 / x86_64）+ 状态页启动状态 + 全局作用域
多架构：
- build.py 增加 armeabi-v7a / x86 目标；customize.sh 按 $ARCH 从 libs/<arch> 解包
- setup_cargo_config.py 生成四个 target 的 linker/bindgen 配置（arm32 用
  armv7a-linux-androideabi clang，sysroot include 目录是 arm-linux-androideabi）
- scripts/pack_combined.py：把四个按 ABI 构建的 zip 合成单个发布 zip（同一份
  SUPPORTED_ABIS 声明全部架构），模块管理器只认一个 zipUrl 所以发布件必须是合并的
- 32 位移植（这些是首次编 32 位时暴露的真实问题）：
  * keymint: tv_sec/tv_nsec 在 32 位是 i32（clock.rs / utils.rs）→ 先 as i64 再运算
  * injector sys.rs: libc 只为 64 位 Android 提供 ptrace 寄存器结构体，32 位就地按
    内核布局定义（arm: uregs[18]；x86: ebx..xss 17 字段）
  * injector fd.rs: st_mode/S_IFMT/S_IFCHR 类型随 ABI 变 → 统一按 mode_t 比较
  * injector payload_fd.rs: msghdr.msg_namelen 32 位是 i32 → as _ 让编译器推断
  * 构建环境：NDK 里 32 位 x86 运行时名为 libclang_rt.builtins-i686-android.a，而
    lsplt 的 build 脚本按 x86 拼库名，需在 NDK 的 clang lib 目录建别名才能链接 x86

A 端 WebUI：
- 远程配置新增「全局作用域」「查看在线 B 端设备 ID」；状态页入口点击后立即跳转，
  页面内提示 4 秒，并同时发一条系统通知（cmd notification post），跳到浏览器后
  在下拉通知栏仍能看到说明
- 提示/按钮文案补齐到全部 22 个语言（此前只有 en/zh-CN，其它语言渲染为空白）
- 保存时保留它不认识的配置键（避免手写的 soter_* 被整份重写抹掉）

发布件：a-side/build/ommega-a-release-1.4.0.zip（102 项，内含四套 lib + 7 个内核系列
pathmask ko，全部 .sha256 自校验通过；module.prop version=1.4.0-179b829 / versionCode=19）
update.json 同步到 1.4.0 / 19，模块管理器据此提示更新

+++ a-side/source/src/utils.rs
@@ -12,7 +12,8 @@ pub fn get_current_time_in_milliseconds() -> i64 {
     // SAFETY: The pointer is valid because it comes from a reference, and clock_gettime doesn't
     // retain it beyond the call.
     unsafe { libc::clock_gettime(libc::CLOCK_BOOTTIME, &mut current_time) };
-    current_time.tv_sec * 1000 + (current_time.tv_nsec / 1_000_000)
+    // `tv_sec`/`tv_nsec` are i32 on 32-bit ABIs (armv7/i686), i64 on 64-bit.
+    (current_time.tv_sec as i64) * 1000 + (current_time.tv_nsec as i64) / 1_000_000
 }
 
 pub trait ParcelExt {
